Remove "further options" separator
OSBuild is certainly neither "not actively maintained" nor "more limited".
Also Lorax, while having some limitations, is a powerfull tool that is
actively used. In general, this split is very very subjective. E.g.
"packer" is certainly "more limited", but it was listed above the
separator. The split is not useful, drop it.